+++ packages/sdk/cms/src/services/menuService.ts
@@ -1,6 +1,7 @@
 import type { MenuVariables } from '@geins/types';
 import { BaseApiService, logWrite } from '@geins/core';
 import { queries } from '../graphql';
+import { valueFromAST } from 'graphql';
 export class MenuService extends BaseApiService {
   private async generateVars(variables: MenuVariables) {
     if (!variables.menuLocationId) {
@@ -41,19 +42,19 @@ export class MenuService extends BaseApiService {
   }
 
   protected parseMenuItem(item: any): any {
+    console.log('item', item);
     return {
       id: item.id,
       label: item.label,
       title: item.title,
       canonicalUrl: item.canonicalUrl,
+      value: item.value,
       type: item.type,
       order: item.order,
       open: item.open,
       hidden: item.hidden,
       targetBlank: item.targetBlank,
-      children: item.children
-        ? item.children.map((child: any) => this.parseMenuItem(child))
-        : [],
+      children: item.children ? item.children.map((child: any) => this.parseMenuItem(child)) : [],
     };
   }
 }