+++ packages/sdk/core/__tests__/Graphql.test.ts
@@ -14,7 +14,7 @@ describe('GeinsCMS', () => {
   afterEach(() => {
     jest.clearAllMocks();
   });
-
+  /*
   it('should get result from custom query', async () => {
     const MY_QUERY = gql`
       query getChannels {
@@ -39,5 +39,82 @@ describe('GeinsCMS', () => {
     expect(result).toBeDefined();
     expect(result!.channels).toBeDefined();
     expect(result!.channels.length).toBeGreaterThan(0);
+  });
+
+  it('should get result from custom string query', async () => {
+    const MY_QUERY = `
+      query getChannels {
+        channels {
+          name
+          url
+        }
+      }
+    `;
+
+    interface Channels {
+      channels: {
+        name: string;
+        url: string;
+      }[];
+    }
+
+    graphqlClient.console_log = true;
+
+    const result = await graphqlClient.query<Channels>({
+      queryAsString: MY_QUERY,
+    });
+
+    expect(result).toBeDefined();
+    expect(result!.channels).toBeDefined();
+    expect(result!.channels.length).toBeGreaterThan(0);
+  }); */
+
+  it('should add standard vars is if present in query', async () => {
+    const MY_QUERY = `
+    query listPageInfo(
+    $url: String!
+    $channelId: String
+    $languageId: String
+    $marketId: String
+    ) {
+      listPageInfo(
+        url: $url
+        channelId: $channelId
+        languageId: $languageId
+        marketId: $marketId
+      ) {
+            id
+            alias
+            canonicalUrl
+            primaryImage
+            name
+            primaryDescription
+            secondaryDescription
+            hideTitle
+            hideDescription
+            logo
+            meta {
+               title
+                description
+            }
+            subCategories {
+              name
+              alias
+              canonicalUrl
+            }
+      }
+    }
+    `;
+
+    graphqlClient.log_to_console = true;
+
+    const result = await graphqlClient.query<any>({
+      queryAsString: MY_QUERY,
+      variables: {
+        url: 'home-interior',
+      },
+    });
+
+    expect(result).toBeDefined();
   });
 });