+++ packages/config/eslint-config/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @geins/eslint-config
 
+## 0.1.6-canary-848c93a8f7473b2be21671e5c82eb3dca0788613
+
+### Patch Changes
+
+- wip
+
+## 0.1.6-canary-rc.{version}
+
+### Patch Changes
+
+- wip
+
+## 0.1.6-canary
+
+### Patch Changes
+
+- d666708: added settingvars to GraphQLService
+- 2158dcd: Possiblity to use string as query for the merchantAPI
+- 5f86eeb: ci fixes
+
 ## 0.1.5
 
 ### Patch Changes