+++ packages/sdk/crm/CHANGELOG.md
@@ -1,5 +1,73 @@
 # @geins/crm
 
+## 0.3.5-canary-848c93a8f7473b2be21671e5c82eb3dca0788613
+
+### Patch Changes
+
+- wip
+- Updated dependencies
+  - @geins/core@0.3.5-canary-848c93a8f7473b2be21671e5c82eb3dca0788613
+
+## 0.3.5-canary-rc.{version}
+
+### Patch Changes
+
+- wip
+- Updated dependencies
+  - @geins/core@0.3.5-canary-rc.{version}
+
+## 0.3.5-canary-DATETIMESTAMP
+
+### Patch Changes
+
+- wip
+- Updated dependencies
+  - @geins/core@0.3.5-canary-DATETIMESTAMP
+
+## 0.3.5-canary
+
+### Patch Changes
+
+- wip
+- Updated dependencies
+  - @geins/core@0.3.5-canary
+
+## 0.3.5-canary
+
+### Patch Changes
+
+- wip
+- Updated dependencies
+  - @geins/core@0.3.5-canary
+
+## 0.3.5-canary
+
+### Patch Changes
+
+- canary bump
+- Updated dependencies
+  - @geins/core@0.3.5-canary
+
+## 0.3.5-canary
+
+### Patch Changes
+
+- update canary
+- Updated dependencies
+  - @geins/core@0.3.5-canary
+
+## 0.3.5-canary
+
+### Patch Changes
+
+- d666708: added settingvars to GraphQLService
+- 2158dcd: Possiblity to use string as query for the merchantAPI
+- 5f86eeb: ci fixes
+- Updated dependencies [d666708]
+- Updated dependencies [2158dcd]
+- Updated dependencies [5f86eeb]
+  - @geins/core@0.3.5-canary
+
 ## 0.3.4
 
 ### Patch Changes