+++ packages/sdk/types/CHANGELOG.md
@@ -1,5 +1,54 @@
 # @geins/types
 
+## 0.3.5-canary-848c93a8f7473b2be21671e5c82eb3dca0788613
+
+### Patch Changes
+
+- wip
+
+## 0.3.5-canary-rc.{version}
+
+### Patch Changes
+
+- wip
+
+## 0.3.5-canary-DATETIMESTAMP
+
+### Patch Changes
+
+- wip
+
+## 0.3.5-canary
+
+### Patch Changes
+
+- wip
+
+## 0.3.5-canary
+
+### Patch Changes
+
+- wip
+
+## 0.3.5-canary
+
+### Patch Changes
+
+- canary bump
+
+## 0.3.5-canary
+
+### Patch Changes
+
+- update canary
+
+## 0.3.5-canary
+
+### Patch Changes
+
+- d666708: added settingvars to GraphQLService
+- 2158dcd: Possiblity to use string as query for the merchantAPI
+
 ## 0.3.4
 
 ### Patch Changes